src/services/streamlit_app/requirements.txt
streamlit
streamlit-image-comparison
requests
pillow
